requirements.txt
requests
Pillow
pygame